+++ ThirdPartyAdapters/ironsource/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## IronSource Android Mediation Adapter Changelog
 
+#### Next Version
+- Added support for forwarding the tagForUnderAgeOfConsent Google Mobile Ads SDK parameter to the ironSource SDK.
+
 #### Version 9.3.0.0 (In progress)
 
 #### Version 9.2.0.0